+++ CHANGELOG.md
@@ -2,6 +2,20 @@
 
 This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html)
 
+## Unreleased (2024-)
+
+### Breaking changes
+
+Changed external variables to be requested in runtime instead of set beforehand.
+This means now there are two new methods to define the callbacks (`on_external_variable_fetch`, `on_external_variable_update`) and
+the external variable methods `set_external_variable` and `get_external_variable` were removed.
+
+### Added
+
+- External variable fetching/updating callbacks
+- Editor player now supports external variables
+
+
 ## 4.0.0 (2024-02-14)
 
 ### Breaking changes